refactor: extract st0x-raindex crate with the Raindex trait and domain types
