+++ src/onchain/raindex.rs
@@ -11,67 +11,22 @@
 //! OrderBook V6 uses a custom float format (B256) for amounts. All conversions between
 //! standard fixed-point amounts (U256) and the float format MUST use rain-math-float.
 
-use alloy::primitives::{Address, B256, TxHash, U256, address};
+use alloy::primitives::{Address, TxHash, U256};
 use alloy::providers::Provider;
 use alloy::rpc::types::{Filter, TransactionReceipt};
 use alloy::sol_types::SolEvent;
-use alloy::transports::{RpcError, TransportErrorKind};
 use async_trait::async_trait;
 use rain_math_float::Float;
 use tracing::{debug, info};
 
-use st0x_evm::{Evm, EvmError, IntoErrorRegistry, OpenChainErrorRegistry, Wallet};
+use st0x_evm::{Evm, IntoErrorRegistry, OpenChainErrorRegistry, Wallet};
 use st0x_execution::FractionalShares;
 use st0x_finance::Usdc;
+pub(crate) use st0x_raindex::{Raindex, RaindexError, RaindexVaultId, USDC_BASE};
 
 use crate::bindings::{IERC20, IOrderBookV6};
 
-pub(crate) const USDC_BASE: Address = address!("0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913");
 const USDC_DECIMALS: u8 = 6;
-
-/// Vault identifier for Rain OrderBook vaults.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(crate) struct RaindexVaultId(pub(crate) B256);
-
-#[derive(Debug, thiserror::Error)]
-pub(crate) enum RaindexError {
-    #[error("EVM error: {0}")]
-    Evm(#[from] EvmError),
-    #[error("Contract error: {0}")]
-    Contract(#[from] alloy::contract::Error),
-    #[error("Float error: {0}")]
-    Float(#[from] rain_math_float::FloatError),
-    #[error("Amount cannot be zero")]
-    ZeroAmount,
-    #[error("RPC transport error: {0}")]
-    RpcTransport(#[from] RpcError<TransportErrorKind>),
-    #[error("ABI decode error: {0}")]
-    SolType(#[from] alloy::sol_types::Error),
-    /// A withdrawal scan could not confirm presence or absence: the queried node
-    /// is not confirmations-deep past `from_block`, so an empty result may be RPC
-    /// lag rather than a true absence. Retryable -- the caller must NOT re-execute
-    /// the irreversible withdraw on this.
-    #[error("withdrawal scan inconclusive: node not caught up past block {from_block}")]
-    ScanInconclusive { from_block: u64 },
-}
-
-impl RaindexError {
-    /// `true` if this error reports that a submitted transaction was dropped from
-    /// the mempool and will never mine -- a terminal failure, distinct from a
-    /// still-pending transaction that simply has not confirmed yet.
-    pub(crate) fn is_transaction_dropped(&self) -> bool {
-        match self {
-            Self::Evm(EvmError::TransactionDropped { .. }) => true,
-            Self::Evm(_)
-            | Self::Contract(_)
-            | Self::Float(_)
-            | Self::ZeroAmount
-            | Self::RpcTransport(_)
-            | Self::SolType(_)
-            | Self::ScanInconclusive { .. } => false,
-        }
-    }
-}
 
 /// Number of `eth_getLogs` scans that must agree the effect is absent before a
 /// resume re-executes an irreversible withdraw. Defends against a single
@@ -411,56 +366,6 @@ impl<E: Evm> RaindexService<E> {
     }
 }
 
-/// Abstraction for Raindex (Rain OrderBook) operations.
-///
-/// This trait abstracts deposit and withdraw operations for Raindex,
-/// allowing different implementations (real service, mock) to be used interchangeably.
-#[async_trait]
-pub(crate) trait Raindex: Send + Sync {
-    /// Withdraws tokens from a Rain OrderBook vault (atomic submit + confirm).
-    async fn withdraw(
-        &self,
-        token: Address,
-        vault_id: RaindexVaultId,
-        target_amount: U256,
-        decimals: u8,
-    ) -> Result<TxHash, RaindexError>;
-
-    /// Submit a vault deposit without waiting for confirmation.
-    ///
-    /// Handles approval if needed, submits the deposit4 transaction,
-    /// and returns the tx hash immediately. Use
-    /// [`confirm_tx`](Raindex::confirm_tx) to wait for confirmation.
-    async fn submit_deposit(
-        &self,
-        token: Address,
-        vault_id: RaindexVaultId,
-        amount: U256,
-        decimals: u8,
-    ) -> Result<TxHash, RaindexError>;
-
-    /// Submit a vault withdrawal without waiting for confirmation.
-    ///
-    /// Returns the tx hash immediately. Use
-    /// [`confirm_tx`](Raindex::confirm_tx) to wait for confirmation.
-    async fn submit_withdraw(
-        &self,
-        token: Address,
-        vault_id: RaindexVaultId,
-        target_amount: U256,
-        decimals: u8,
-    ) -> Result<TxHash, RaindexError>;
-
-    /// Wait for a previously submitted transaction to be confirmed.
-    async fn confirm_tx(&self, tx_hash: TxHash) -> Result<(), RaindexError> {
-        self.confirm_tx_receipt(tx_hash).await.map(|_| ())
-    }
-
-    /// Wait for a previously submitted transaction to be confirmed and return the receipt.
-    async fn confirm_tx_receipt(&self, tx_hash: TxHash)
-    -> Result<TransactionReceipt, RaindexError>;
-}
-
 #[async_trait]
 impl<W: Wallet> Raindex for RaindexService<W> {
     async fn withdraw(
@@ -559,7 +464,7 @@ mod tests {
     use alloy::network::Ethereum;
     use alloy::node_bindings::{Anvil, AnvilInstance};
     use alloy::primitives::Log as PrimitiveLog;
-    use alloy::primitives::{B256, b256};
+    use alloy::primitives::{B256, address, b256};
     use alloy::providers::Provider;
     use alloy::providers::fillers::{
         BlobGasFiller, ChainIdFiller, FillProvider, GasFiller, JoinFill, NonceFiller,
@@ -573,6 +478,7 @@ mod tests {
     use alloy::providers::mock::Asserter;
     use serde_json::json;
 
+    use st0x_evm::EvmError;
     use st0x_evm::NoOpErrorRegistry;
     use st0x_evm::ReadOnlyEvm;
     use st0x_evm::Wallet;
